SMTH-6: Attack-defend cycle implementation
When an actor processes an action, the action acts as a mediator between the actor and its targets. We get and attackConfig from the actor and allow it's target to process it via the processAttack API. From there the target's behviourCtrl takes care of processing the attack and applying the changes to its actor as needed. Once complete, an AttackResult is returned through an out reference so the action can then route the result data back to its actor, or elsewhere.

- Actors can now process attacks against and take damage
- Added additional stats to actors for simulation testing
- Actors now write to an AttackResult object that holds data about what happened while processing an attack